Move the cli content in a dedicated file
